requirements.txt
streamlit
openai==1.10.0
langchain-community==0.0.16
langchain-openai==0.0.5
langchain-core==0.1.16
tiktoken
faiss-cpu
langchain==0.1.4
python-dotenv
pysqlite3-binary
chromadb==0.4.22